Fixed the name of the repo
Because the Go dependencies are case-sensitive, the folder's name must match with the repo's name. Otherwise, `dep` will put a copy of the repo in the `vendor` folder and will use it for building rather than the actual codebase.

Spent 4+ hours to realize what wrong was happening, and this statement in the installation step turned out to be the culprit. Fixed it for everyone in future.