Update README.md
For trending feeds we use total number of participants. Right now you would need to fetch a claim, iterate through claim arguments and then iterate over staking lists for each argument. I though since we're keeping the `Total(Backing/Challenge)Stake` on the claim, we might as well include the `TotalParticipants`.

Also, I'm not sure if we want to follow this convention or not, but we could suffix the word `Stake` whenever we refer to sdk.Coin's.